Hotel/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
cos-python-sdk-v5==1.7.7
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20181111
Django==3.0.2
django-crispy-forms==1.8.1
django-formtools==2.2
django-import-export==2.0
django-reversion==3.0.5
django-simpleui==3.9.1
et-xmlfile==1.0.1
freetype-py==2.1.0.post1
future==0.18.2
HPSocket==0.9.3
httplib2==0.15.0
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
numpy==1.18.1
odfpy==1.4.0
opencv-python==4.1.2.30
openpyxl==3.0.2
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
qcloudsms-py==0.1.4
redis==3.3.11
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.7
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
